requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-heroku==0.3.1
psycopg2==2.9.5
sqlparse==0.4.3
whitenoise==6.4.0
cryptography==39.0.1
django-appconf==1.0.5
django-cors-headers==3.13.0
django-cryptography==1.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.9.0
gunicorn==20.1.0
iniconfig==1.1.1
mpld3==0.5.1
mypy-extensions==0.4.3
openpyxl==3.1.1
packaging==20.9
pipenv==2022.12.19
